requirements.txt
########Python==3.8.10#########
Django==4.2.20
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
easyocr==1.7.2
numpy==1.24.1
open-clip-torch==2.30.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
dotenv==0.9.9
whitenoise==6.7.0
social-auth-app-django==5.4.2
requests-oauthlib==2.0.0
cryptography==42.0.5
django-storages==1.14.5
pytesseract==0.3.13
django-allauth==65.5.0
boto3==1.37.13
stripe==12.0.0
twilio==9.5.2
transformers==4.46.3
fuzzywuzzy
python-Levenshtein==0.21.0
bs4
pytrends
geopy
django-cors-headers==4.4.0
openai
azure-cognitiveservices-vision-computervision
python-jose
firebase-admin==6.5.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
psycopg2-binary>=2.9
gunicorn